ringct: fix a warning on int type

+++ src/math/ringct/functional/rctOps.cpp
@@ -134,7 +134,7 @@ namespace rct {
 
     crypto::ec_scalar r = crypto::int_to_scalar(x);
 
-    for (int i = 0; i < sizeof(uint64_t); ++i)
+    for (size_t i = 0; i < sizeof(uint64_t); ++i)
       r.data[i] ^= h.data[i];
 
     return crypto::scalar_to_int(r);
